Build now working and removed or translated alla swe comments

+++ src/app/lib/delete-account.ts
@@ -1,12 +1,17 @@
-import { UserLogin } from "dexie-cloud-addon"
-import { db } from "../db/db"
-import { AppRouterInstance } from "next/dist/shared/lib/app-router-context.shared-runtime"
+import { UserLogin } from 'dexie-cloud-addon'
+import { db } from '../db/db'
+import { AppRouterInstance } from 'next/dist/shared/lib/app-router-context.shared-runtime'
 
 export const deleteUserAccount = async (
-  user: UserLogin,
-  router: AppRouterInstance
+  user:
+    | UserLogin
+    | {
+        userId: string
+        email: string
+      },
+  router: AppRouterInstance,
 ) => {
-  if (!user?.userId) return // Safety check
+  if (!user?.userId) return
 
   const confirmed =
     confirm(`Are you sure you want to delete your user completely along all stored data for ${user?.userId}?
@@ -18,22 +23,26 @@ Private data will be deleted. Shared data will not be deleted. This action canno
   try {
     const url = `${db.cloud.options?.databaseUrl}/users/${user.userId}`
 
-    const options = {
-      url,
-      method: "DELETE",
-      headers: {
-        Authorization: `Bearer ${user.accessToken}`,
-      },
-    }
+    // Check if the user has an access token
+    if ('accessToken' in user && user.accessToken) {
+      const options = {
+        method: 'DELETE',
+        headers: {
+          Authorization: `Bearer ${user.accessToken}`,
+        },
+      }
 
-    await fetch(options)
-      .catch((error) => {
-        console.error("Error deleting user", error)
-      })
-      .then(() => {
-        router.push("/logout")
-      })
+      await fetch(url, options)
+        .catch((error) => {
+          console.error('Error deleting user', error)
+        })
+        .then(() => {
+          router.push('/logout')
+        })
+    } else {
+      console.error('User does not have an access token.')
+    }
   } catch (error) {
-    console.error("Error deleting user", error)
+    console.error('Error deleting user', error)
   }
 }
